backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
